job-optimizer/requirements.txt
streamlit==1.29.0
pandas==2.1.4
groq==0.4.1
langchain==0.1.0
langchain-groq==0.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1
plotly==5.18.0
langgraph==0.0.15
typing_extensions==4.9.0

